server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.12.5
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.28.1
pydantic-settings==2.1.0
langgraph==1.0.5
langchain-openai==1.1.6
langchain-community==0.4.1
langchain-core==1.2.6
unstructured[pdf]==0.18.31
xhtml2pdf==0.2.17
pymysql==1.1.2
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.45
tiktoken==0.12.0
email-validator==2.3.0
redis==5.0.1
sse-starlette==2.0.0
